Switch to int64 instead of int

+++ model/build_version.go
@@ -5,8 +5,8 @@ type BuildVersions []*BuildVersion
 
 // BuildVersion represents a build version model definition.
 type BuildVersion struct {
-	BuildID   int      `json:"build_id" sql:"index"`
+	BuildID   int64    `json:"build_id" sql:"index"`
 	Build     *Build   `json:"build,omitempty"`
-	VersionID int      `json:"version_id" sql:"index"`
+	VersionID int64    `json:"version_id" sql:"index"`
 	Version   *Version `json:"version,omitempty"`
 }
