+++ model/pack_icon.go
@@ -16,8 +16,8 @@ type PackIcons []*PackIcon
 
 // PackIcon represents a pack icon model definition.
 type PackIcon struct {
-	ID          int              `json:"-" gorm:"primary_key"`
-	PackID      int              `json:"-" sql:"index"`
+	ID          int64            `json:"-" gorm:"primary_key"`
+	PackID      int64            `json:"-" sql:"index"`
 	Pack        *Pack            `json:"-"`
 	Slug        string           `json:"slug" sql:"unique_index"`
 	ContentType string           `json:"content_type"`
