+++ model/user_pack.go
@@ -5,9 +5,9 @@ type UserPacks []*UserPack
 
 // UserPack represents a user pack model definition.
 type UserPack struct {
-	UserID int    `json:"user_id" sql:"index"`
+	UserID int64  `json:"user_id" sql:"index"`
 	User   *User  `json:"user,omitempty"`
-	PackID int    `json:"pack_id" sql:"index"`
+	PackID int64  `json:"pack_id" sql:"index"`
 	Pack   *Pack  `json:"pack,omitempty"`
 	Perm   string `json:"perm,omitempty"`
 }
