+++ model/version.go
@@ -14,10 +14,10 @@ type Versions []*Version
 
 // Version represents a version model definition.
 type Version struct {
-	ID        int          `json:"id" gorm:"primary_key"`
+	ID        int64        `json:"id" gorm:"primary_key"`
 	File      *VersionFile `json:"file,omitempty"`
 	Mod       *Mod         `json:"mod,omitempty"`
-	ModID     int          `json:"mod_id" sql:"index"`
+	ModID     int64        `json:"mod_id" sql:"index"`
 	Slug      string       `json:"slug"`
 	Name      string       `json:"name"`
 	CreatedAt time.Time    `json:"created_at"`
